feat(portal): move page fetching functions to common lib
